use cpuid.py
